% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_plot.R
\name{beta_plot}
\alias{beta_plot}
\title{plot estimated single index beta value}
\usage{
beta_plot(beta_est)
}
\arguments{
\item{beta_est}{A data set for beta plot}
}
\value{
beta value plot
}
\description{
plot estimated single index beta value
}
\examples{
beta_est <- beta_est
beta_plot(beta_est=beta_est)
}
