% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.trans.R
\name{covariate.trans}
\alias{covariate.trans}
\title{confounder vector Z transformation}
\usage{
covariate.trans(Z_continuous, Z_discrete, data)
}
\arguments{
\item{Z_continuous}{A character name vector for continuous confounder}

\item{Z_discrete}{A character name vector for discrete confounder}

\item{data}{A data set}
}
\value{
updated confounder vector and data set.
}
\description{
confounder vector Z transformation
}
\examples{
\dontrun{
sum("a")
}
}
