% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_trans.R
\name{covariate_trans}
\alias{covariate_trans}
\title{covariate vector Z transformation}
\usage{
covariate_trans(Z_continuous, Z_discrete, data)
}
\arguments{
\item{Z_continuous}{A character vector for continuous covariates}

\item{Z_discrete}{A character vector for discrete covariates}

\item{data}{A data set}
}
\value{
updated covariate vector and data set.
}
\description{
covariate vector Z transformation
}
\examples{
\dontrun{
sum("a")
}
}
