% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.interaction.plot.R
\name{exp.interaction.plot}
\alias{exp.interaction.plot}
\title{plot interaction effect of two exposures}
\usage{
\method{exp}{interaction.plot}(fit, data, exp_1, exp_2)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi.lr.v1'}

\item{data}{Original data set}

\item{exp_1}{exposure name hoping to be checked}

\item{exp_2}{exposure name hoping to be checked}
}
\value{
plot of interaction effect of two exposures with others at average level 0
}
\description{
plot interaction effect of two exposures
}
\examples{
\dontrun{
sum("a")
}
}
