% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.main.plot.R
\name{exp.main.plot}
\alias{exp.main.plot}
\title{plot single exposure's main effect}
\usage{
\method{exp}{main.plot}(fit, data, exp_name)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi.lr.v1'}

\item{data}{Original data set}

\item{exp_name}{exposure name hoping to be plotted}
}
\value{
plot of main effect with other exposures at average level 0
}
\description{
plot single exposure's main effect
}
\examples{
\dontrun{
sum("a")
}
}
