% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interquartile_interquartile_plot.R
\name{interquartile_interquartile_plot}
\alias{interquartile_interquartile_plot}
\title{plot interquartile effect of specific exposure based on quartile of other exposures}
\usage{
interquartile_interquartile_plot(fit, data)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}
}
\value{
plot interquartile effect of specific interquartile exposure
}
\description{
plot interquartile effect of specific exposure based on quartile of other exposures
}
\examples{
\dontrun{
sum("a")
}
}
