% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interquartile_quartile_plot.R
\name{interquartile_quartile_plot}
\alias{interquartile_quartile_plot}
\title{plot interquartile effect of specific exposure based on quartile of other exposures}
\usage{
interquartile_quartile_plot(fit, data, X)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}
}
\value{
plot of main effect with other exposures at average level 0
}
\description{
plot interquartile effect of specific exposure based on quartile of other exposures
}
\examples{
\dontrun{
sum("a")
}
}
