% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 800 rows and 28 columns.
}
\usage{
nhanes
}
\description{
This is data to be included in my package
}
\references{
\url{https://www.cdc.gov/nchs/nhanes/index.htm}
}
\author{
Yuyan Wang \email{yuyan.wang@nyumc.org}
}
\keyword{datasets}
