% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsi.lr.v1.R
\name{plsi.lr.v1}
\alias{plsi.lr.v1}
\title{Partial linear single index linear regression for scalar outcome}
\usage{
plsi.lr.v1(
  data,
  Y.name,
  X.name,
  Z.name,
  spline.num,
  spline.degree,
  initial.random.num
)
}
\arguments{
\item{data}{A data set}

\item{Y.name}{Variable name for scalar outcome}

\item{X.name}{Variable name vector for exposures}

\item{Z.name}{Variable name vector for confounders}

\item{spline.num}{A number representing the degree of freedom of B-spline basis for link function}

\item{spline.degree}{A number representing the degree of the piece-wise polynomial of B-spline basis for link function}

\item{initial.random.num}{A number representing the number of random initials used in the function}
}
\value{
A list of model estimation and prediction results
}
\description{
Partial linear single index linear regression for scalar outcome
}
\examples{
\dontrun{
sum("a")
}
}
