% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsi_lr_v1.R
\name{plsi_lr_v1}
\alias{plsi_lr_v1}
\title{version 1: partial linear single index linear regression for scalar outcome}
\usage{
plsi_lr_v1(data, Y, X, Z, spline_num, spline_degree, initial_random_num)
}
\arguments{
\item{data}{A data set}

\item{Y}{Character for scalar outcome}

\item{X}{A character vector for exposures}

\item{Z}{A character vector for covariates}

\item{spline_num}{A number representing the degree of freedom of B-spline basis for link function}

\item{spline_degree}{A number representing the degree of the piece-wise polynomial of B-spline basis for link function}

\item{initial_random_num}{A number representing the number of random initials used in the function}
}
\value{
A list of model estimation and prediction results
}
\description{
version 1: partial linear single index linear regression for scalar outcome
}
\examples{
data <- nhanes
Y <- "triglyceride"
X <- c("a1.trans.b.carotene","a2.cis.b.carotene","a3.Retinyl.palmitate")
Z <- "age"
spline_num <- 5
spline_degree <- 3
initial_random_num <- 5
plsi_lr_v1(data=data, Y=Y, X=X, Z=Z, spline_num=spline_num, spline_degree=spline_degree, initial_random_num=initial_random_num)
}
