% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsi_qr_v1.R
\name{plsi_qr_v1}
\alias{plsi_qr_v1}
\title{version 1: partial linear single index quantile regression for scalar outcome}
\usage{
plsi_qr_v1(data, Y, X, Z, spline_num, spline_degree, initial_random_num, tau)
}
\arguments{
\item{data}{A data set}

\item{Y}{Character for scalar outcome}

\item{X}{A character vector for exposures}

\item{Z}{A character vector for covariates}

\item{spline_num}{A number representing the degree of freedom of B-spline basis for link function}

\item{spline_degree}{A number representing the degree of the piece-wise polynomial of B-spline basis for link function}

\item{initial_random_num}{A number representing the number of random initials used in the function}

\item{tau}{The modeling quantile level}
}
\value{
A list of model estimation and prediction results
}
\description{
version 1: partial linear single index quantile regression for scalar outcome
}
\examples{
\dontrun{
sum("a")
}
}
