% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_interaction_plot.R
\name{quantile_interaction_plot}
\alias{quantile_interaction_plot}
\title{plot interaction effect based on quantile of exposures}
\usage{
quantile_interaction_plot(fit, data, exp_1, exp_2)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}

\item{exp_1}{exposure name hoping to check}

\item{exp_2}{exposure name hoping to check}
}
\value{
plot of main effect with other exposures at average level 0
}
\description{
plot interaction effect based on quantile of exposures
}
\examples{
\dontrun{
sum("a")
}
}
