% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_main_plot.R
\name{quantile_main_plot}
\alias{quantile_main_plot}
\title{plot main effect based on quantile of exposures}
\usage{
quantile_main_plot(fit, data, exp_name)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}

\item{exp_name}{exposure name hoping to check}
}
\value{
plot of main effect with other exposures at average level 0
}
\description{
plot main effect based on quantile of exposures
}
\examples{
\dontrun{
sum("a")
}
}
