% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_overall_plot.R
\name{quantile_overall_plot}
\alias{quantile_overall_plot}
\title{plot overall effect based on quantile of exposures}
\usage{
quantile_overall_plot(fit, data)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}
}
\value{
plot of predicted outcome based on quantile of exposures
}
\description{
plot overall effect based on quantile of exposures
}
\examples{
\dontrun{
sum("a")
}
}
