% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_order.R
\name{re_order}
\alias{re_order}
\title{change the order of the exposure vector}
\usage{
re_order(X, Y, data)
}
\arguments{
\item{X}{A character vector for exposure mixture}

\item{Y}{A character for outcome}

\item{data}{A data set}
}
\value{
Re-ordered character vector for exposure mixture.
}
\description{
change the order of the exposure vector
}
\examples{
X <- c("a1.trans.b.carotene","a2.cis.b.carotene","a3.Retinyl.palmitate")
Y <- "triglyceride"
data <- nahanes
re_order(X = X, Y = Y, data = dat)
}
