% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.coef.plot.R
\name{si.coef.plot}
\alias{si.coef.plot}
\title{plot estimated single index coefficients}
\usage{
si.coef.plot(si.coef.est)
}
\arguments{
\item{si.coef.est}{A data set of estimated single index coefficients}
}
\value{
single index coefficient plot
}
\description{
plot estimated single index coefficients
}
\examples{
\donttest{
# example to plot estimated single index function -- continuous outcome
data(nhanes.new)
data <- nhanes.new

Y.name <- "log.triglyceride"
X.name <- c("X1_trans.b.carotene", "X2_retinol", "X3_g.tocopherol", "X4_a.tocopherol",
            "X5_PCB99", "X6_PCB156", "X7_PCB206",
            "X8_3.3.4.4.5.pncb", "X9_1.2.3.4.7.8.hxcdf", "X10_2.3.4.6.7.8.hxcdf")
Z.name <- c("AGE.c", "SEX.Female", "RACE.NH.Black",
           "RACE.MexicanAmerican", "RACE.OtherRace", "RACE.Hispanic" )

k <- 10
bs <- "cr"
initial.random.num <- 1
seed = 2026

model_lr_auto <- plsi.lr.auto(data = data, Y.name = Y.name, X.name = X.name, Z.name = Z.name,
                      k = k, bs = bs, initial.random.num = initial.random.num, seed = seed)

# plot estimated single index coefficients
si.coef.plot(model_lr_auto$si.coefficient)

# check estimated single index coefficients
model_lr_auto$si.coefficient

# example to plot estimated single index function -- binary outcome
data$high.triglyceride <- as.numeric(
  data$log.triglyceride > stats::quantile(data$log.triglyceride, 2 / 3)
)
model_logistic_auto <- plsi.logistic.auto(data = data, Y.name = "high.triglyceride",
                      X.name = X.name, Z.name = Z.name,
                      k = k, bs = bs, initial.random.num = initial.random.num, seed = seed)

# plot estimated single index coefficients
si.coef.plot(model_logistic_auto$si.coefficient)

# check estimated single index coefficients
model_logistic_auto$si.coefficient


# example to plot estimated single index function -- count outcome
set.seed(2026)
beta_true <- c(0.30, -0.20, 0.10, 0.40, -0.30, 0.20, -0.10, 0.25, -0.15, 0.35)
beta_true <- beta_true / sqrt(sum(beta_true^2))
x_std <- scale(data[, X.name])
single_index_true <- as.vector(x_std \%*\% beta_true)
log_rate <- 0.3 + 0.4 * sin(single_index_true) + 0.05 * data$AGE.c
data$n.events <- stats::rpois(nrow(data), lambda = exp(log_rate))
model_log_auto <- plsi.log.auto(data = data, Y.name = "n.events", X.name = X.name,
                      Z.name = Z.name, family = "nb", k = k, bs = bs,
                      initial.random.num = initial.random.num, seed = seed)
# plot estimated single index coefficients
si.coef.plot(model_log_auto$si.coefficient)

# check estimated single index coefficients
model_log_auto$si.coefficient
}
}
\author{
Yuyan Wang
}
\keyword{coefficients}
\keyword{index}
\keyword{linear}
\keyword{partial}
\keyword{single}
