% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounder.trans.R
\name{confounder.trans}
\alias{confounder.trans}
\title{Transformation for confounder vector Z}
\usage{
confounder.trans(Z_continuous, Z_discrete, data)
}
\arguments{
\item{Z_continuous}{A character name vector for continuous confounders}

\item{Z_discrete}{A character name vector for discrete confounders}

\item{data}{Orginial data set}
}
\value{
Transformed confounder vector and data set ready for further analysis.
}
\description{
Transformation for confounder vector Z
}
\author{
Yuyan Wang
}
\keyword{confounder}
