% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.interaction.plot.R
\name{e.interaction.plot}
\alias{e.interaction.plot}
\title{plot interaction effect of two exposures}
\usage{
e.interaction.plot(fit, data, exp_1, exp_2)
}
\arguments{
\item{fit}{Fitted model from function 'plsi.lr.v1'}

\item{data}{Original data set}

\item{exp_1}{exposure name hoping to be checked}

\item{exp_2}{exposure name hoping to be checked}
}
\value{
plot of interaction effect of two exposures with others at average level
}
\description{
plot interaction effect of two exposures
}
\author{
Yuyan Wang
}
\keyword{exposures.interaction.effect}
