% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.main.plot.R
\name{e.main.plot}
\alias{e.main.plot}
\title{plot single exposure's main effect}
\usage{
e.main.plot(fit, data, exp_name)
}
\arguments{
\item{fit}{Fitted model from function 'plsi.lr.v1'}

\item{data}{Original data set}

\item{exp_name}{exposure name hoping to be plotted}
}
\value{
plot of exposure's main effect with other exposures at average level 0
}
\description{
plot single exposure's main effect
}
\author{
Yuyan Wang
}
\keyword{exposure.main.effect}
