% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interquartile.quartile.plot.R
\name{interquartile.quartile.plot}
\alias{interquartile.quartile.plot}
\title{plot interquartile effect of specific exposure based on quartile of other exposures}
\usage{
interquartile.quartile.plot(fit, data, type = c("linear", "logistic", "log"))
}
\arguments{
\item{fit}{Fitted model from \code{plsi.lr.auto()}, \code{plsi.logistic.auto()},
or \code{plsi.log.auto()}}

\item{data}{Original data set}

\item{type}{Which outcome scale to plot. \code{"linear"} (default) plots the
difference in predicted (continuous) outcome, for a \code{fit} from
\code{plsi.lr.auto()}. \code{"logistic"} plots the difference in predicted
probability, for a \code{fit} from \code{plsi.logistic.auto()}. \code{"log"}
plots the difference in predicted count, for a \code{fit} from
\code{plsi.log.auto()}.}
}
\value{
plot of main interquartile effect of exposure based on quartile of other exposures
}
\description{
plot interquartile effect of specific exposure based on quartile of other exposures
}
\examples{
\donttest{
# example to interquartile effect based on quartile of other exposures -- continuous outcome
data(nhanes.new)
data <- nhanes.new

Y.name <- "log.triglyceride"
X.name <- c("X1_trans.b.carotene", "X2_retinol", "X3_g.tocopherol", "X4_a.tocopherol",
            "X5_PCB99", "X6_PCB156", "X7_PCB206",
            "X8_3.3.4.4.5.pncb", "X9_1.2.3.4.7.8.hxcdf", "X10_2.3.4.6.7.8.hxcdf")
Z.name <- c("AGE.c", "SEX.Female", "RACE.NH.Black",
           "RACE.MexicanAmerican", "RACE.OtherRace", "RACE.Hispanic" )

k <- 10
bs <- "cr"
initial.random.num <- 1
seed = 2026

model_lr_auto <- plsi.lr.auto(data = data, Y.name = Y.name, X.name = X.name, Z.name = Z.name,
                      k = k, bs = bs, initial.random.num = initial.random.num, seed = seed)

# plot interquartile quartile -- difference in predicted (continuous) outcome
interquartile.quartile.plot(model_lr_auto, data, type = "linear")
}
}
\author{
Yuyan Wang
}
\keyword{effect}
\keyword{index}
\keyword{interaction}
\keyword{interquartile}
\keyword{linear}
\keyword{partial}
\keyword{quartile}
\keyword{single}
