% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interquartile.quartile.plot.R
\name{interquartile.quartile.plot}
\alias{interquartile.quartile.plot}
\title{plot interquartile effect of specific exposure based on quartile of other exposures}
\usage{
interquartile.quartile.plot(fit, data)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi.lr.v1'}

\item{data}{Original data set}
}
\value{
plot of main interquartile effect of specific exposure based on quartile of other exposures
}
\description{
plot interquartile effect of specific exposure based on quartile of other exposures
}
\examples{
\dontrun{
sum("a")
}
}
