% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_plot.R
\name{link_plot}
\alias{link_plot}
\title{plot estimated single index link function}
\usage{
link_plot(link_ci)
}
\arguments{
\item{link_ci}{A data set of estimated index, link value, and CIs}
}
\value{
Link function plot
}
\description{
plot estimated single index link function
}
\examples{
\dontrun{
sum("a")
}
}
