% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_plot.R
\name{link_plot}
\alias{link_plot}
\title{plot estimated single index link function}
\usage{
link_plot(index, link, cut)
}
\arguments{
\item{index}{A numeric vector of estimated single index}

\item{link}{A numeric vector of link function value}

\item{cut}{A numeric value of percentile cut in two tails of x-axis}
}
\value{
Link function plot
}
\description{
plot estimated single index link function
}
\examples{
index <- index
link <- link
cut <- 0.02
link_plot(index=index, link=link, cut=cut)
}
