% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.overall.plot.R
\name{mixture.overall.plot}
\alias{mixture.overall.plot}
\title{plot mixture's overall effect based on quantile of exposures}
\usage{
mixture.overall.plot(fit, data)
}
\arguments{
\item{fit}{Fitted model from PLSI function 'plsi_lr_v1'}

\item{data}{Original data set}
}
\value{
plot of predicted outcome based on quantile of exposures
}
\description{
plot mixture's overall effect based on quantile of exposures
}
\examples{
\dontrun{
sum("a")
}
}
