% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.coef.plot.R
\name{si.coef.plot}
\alias{si.coef.plot}
\title{plot estimated single index coefficients}
\usage{
si.coef.plot(si.coef.est)
}
\arguments{
\item{si.coef.est}{A data set of estimated single index coefficients}
}
\value{
single index coefficient plot
}
\description{
plot estimated single index coefficients
}
\author{
Yuyan Wang
}
\keyword{single.index.coefficients}
