% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.fun.plot.R
\name{si.fun.plot}
\alias{si.fun.plot}
\title{plot estimated single index function}
\usage{
si.fun.plot(si.ci)
}
\arguments{
\item{si.ci}{A data set of estimated index, link value, and CIs}
}
\value{
Single index function plot
}
\description{
plot estimated single index function
}
\examples{
\dontrun{
sum("a")
}
}
